+++ src/main/kotlin/net/frozenblock/lib/config/api/client/gui/EntryTypes.kt
@@ -18,8 +18,6 @@
 
 package net.frozenblock.lib.config.api.client.gui
 
-import kotlinx.serialization.Serializable
-
 /**
  * An [Int] wrapper that represents a color.
  * [<P>]
@@ -27,8 +25,53 @@ import kotlinx.serialization.Serializable
  * @param color The [Int] representation of the color
  * @since 1.3.8
  */
-@Serializable
 data class Color(@JvmField val color: Int) {
 
     override fun toString(): String = "Color[$color]"
+
+}
+
+// lists
+
+data class StringList(@JvmField val list: List<String>) {
+
+    override fun toString(): String = "StringList[$list]"
+}
+
+data class IntList(@JvmField val list: List<Int>) {
+
+    override fun toString(): String = "IntList[$list]"
+}
+
+data class LongList(@JvmField val list: List<Long>) {
+
+    override fun toString(): String = "LongList[$list]"
+}
+
+data class FloatList(@JvmField val list: List<Float>) {
+
+    override fun toString(): String = "FloatList[$list]"
+}
+
+data class DoubleList(@JvmField val list: List<Double>) {
+
+    override fun toString(): String = "DoubleList[$list]"
+}
+
+// not lists anymore
+
+data class EnumEntry<T : Enum<?>>(
+    @JvmField val `class`: KClass<T>,
+    @JvmField val value: T
+) {
+
+    override fun toString(): String = "EnumEntry[class=$`class`, value=$value]"
+}
+
+data class SelectorEntry<T>(
+    @JvmField val valuesArray: Array<T>,
+    @JvmField val value: T
+) {
+
+    override fun toString(): String = "SelectorEntry[valuesArray=$valuesArray, value=$value]"
 }