+++ src/main/kotlin/net/frozenblock/lib/config/api/client/gui/Slider.kt
@@ -18,9 +18,6 @@
 
 package net.frozenblock.lib.config.api.client.gui
 
-import kotlinx.serialization.Contextual
-import kotlinx.serialization.Serializable
-
 /**
  * A wrapper of a value, minimum, and maximum that represents a slider.
  * [<P>]
@@ -30,13 +27,9 @@ import kotlinx.serialization.Serializable
  * @param max The maximum [Int] of the slider
  * @since 1.4
  */
-@Serializable
 data class Slider<T>(
-    @Contextual
     val value: Number,
-    @Contextual
     val min: Number,
-    @Contextual
     val max: Number,
     val type: SliderType<T>
 ) {
@@ -44,7 +37,6 @@ data class Slider<T>(
     override fun toString(): String = "Slider[value=$value, min=$min, max=$max]"
 }
 
-@Serializable
 sealed class SliderType<T> {
     data object INT : SliderType<Int>()